Add Headown::Extractor which has logic to extract headers
